requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
click==8.1.7
coverage==7.6.8
distro==1.9.0
fastapi==0.115.6
groq==0.13.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
iniconfig==2.0.0
numpy==2.1.3
packaging==24.2
pluggy==1.5.0
pydantic==2.10.3
pydantic-settings==2.6.1
pydantic_core==2.27.1
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
pytest-mock==3.14.0
python-dotenv==1.0.1
python-multipart==0.0.19
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.32.1
